feat: Game/Room로 Service 분리
